# docs/swagger/swagger.yaml
definitions:
  errors.Error:
    properties:
      detail:
        description: 오류의 상세 설명
        type: string
      instance:
        description: 오류가 발생한 특정 인스턴스를 식별하는 데 사용할 수 있는 URI입니다.
        type: string
      status:
        description: HTTP 상태 코드
        type: integer
      title:
        description: 오류의 짧은 제목
        type: string
    type: object
  presenter.CreateUserRequest:
    properties:
      nickname:
        type: string
      resolution:
        type: string
    required:
    - nickname
    type: object
  presenter.CreateUserResponse:
    properties:
      nickname:
        type: string
      resolution:
        type: string
    type: object
  presenter.FindUserByMeResponse:
    properties:
      nickname:
        type: string
      resolution:
        type: string
    type: object
  presenter.IssueTokenRequest:
    properties:
      nickname:
        type: string
    type: object
  presenter.IssueTokenResponse:
    properties:
      token:
        type: string
    type: object
  presenter.PatchUserByMeRequest:
    properties:
      nickname:
        type: string
      resolution:
        type: string
    type: object
  presenter.ValidateResponse:
    properties:
      nickname:
        type: string
      resolution:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/issue-token:
    post:
      parameters:
      - description: 토큰 발급 요청
        in: body
        name: IssueTokenRequest
        required: true
        schema:
          $ref: '#/definitions/presenter.IssueTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 토큰 발급 성공
          schema:
            $ref: '#/definitions/presenter.IssueTokenResponse'
        "404":
          description: '사용자를 찾을 수 없음 : user not found'
          schema:
            $ref: '#/definitions/errors.Error'
      summary: IssueToken - 토큰 발급 ✅
      tags:
      - auth
  /auth/validate:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 토큰 유효성 검사 성공
          schema:
            $ref: '#/definitions/presenter.ValidateResponse'
        "404":
          description: '사용자를 찾을 수 없음 : user not found'
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Validate - 토큰 유효성 검사 ✅
      tags:
      - auth
  /users:
    post:
      parameters:
      - description: 사용자 생성 요청
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/presenter.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 회원가입에 성공한 사용자 정보
          schema:
            $ref: '#/definitions/presenter.CreateUserResponse'
        "409":
          description: '사용자가 이미 존재함 : user already exists'
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: CreateUser - 사용자 단건 생성 ✅
      tags:
      - users
  /users/me:
    delete:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: '유저를 찾지 못함: user not found'
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: DeleteUserByMe - 사용자(나) 단건 삭제 ✅
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 내 정보 조회에 성공한 사용자 정보
          schema:
            $ref: '#/definitions/presenter.FindUserByMeResponse'
        "404":
          description: '유저를 찾지 못함: user not found'
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: FindUserByMe - 사용자(나) 단건 조회 ✅
      tags:
      - users
    patch:
      parameters:
      - description: 사용자 수정 요청
        in: body
        name: PatchUserByMeRequest
        required: true
        schema:
          $ref: '#/definitions/presenter.PatchUserByMeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: '유저를 찾지 못함: user not found'
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: PatchUserByMe - 사용자(나) 단건 수정 ✅
      tags:
      - users
swagger: "2.0"